+++ types/src/main/java/com/cobber/fta/plugins/address/AddressLineNEN.java
@@ -24,6 +24,8 @@ import com.cobber.fta.AnalyzerContext;
 import com.cobber.fta.Content;
 import com.cobber.fta.Facts;
 import com.cobber.fta.FiniteMap;
+import com.cobber.fta.LogicalType;
+import com.cobber.fta.LogicalTypeFactory;
 import com.cobber.fta.LogicalTypeInfinite;
 import com.cobber.fta.PluginAnalysis;
 import com.cobber.fta.PluginDefinition;
@@ -38,10 +40,15 @@ import com.cobber.fta.token.TokenStreams;
  * Plugin to detect the second line of an Address. (English-language only).
  */
 public abstract class AddressLineNEN extends LogicalTypeInfinite {
-	private PluginLocaleEntry addressLine1Entry;
+	private PluginLocaleEntry previousAddressLineEntry;
 	private PluginLocaleEntry cityEntry;
+	private String country;
+	private LogicalType logicalZip;
+	private LogicalType logicalState;
 	private Set<String> addressMarkers;
 	private final WordProcessor wordProcessor = new WordProcessor().withAdditionalBreakChars("-#").withAdditionalKillChars("'");
+
+	private final String[] previousLine = { "STREET_ADDRESS_EN", "STREET_ADDRESS2_EN", "STREET_ADDRESS3_EN" };
 
 	/**
 	 * Construct a plugin to detect the second line of an Address based on the Plugin Definition.
@@ -51,7 +58,11 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 		super(plugin);
 	}
 
-	protected abstract char getIndicator();
+	protected abstract int getAddressLine();
+
+	private char getAddressLineChar() {
+		return (char)(getAddressLine() + '0');
+	}
 
 	@Override
 	public String nextRandom() {
@@ -72,8 +83,15 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 		super.initialize(analysisConfig);
 
 		addressMarkers = new SingletonSet(new Content("resource", "/reference/en_street_markers.csv")).getMembers();
-		addressLine1Entry = PluginDefinition.findByName("STREET_ADDRESS_EN").getLocaleEntry(locale);
+		previousAddressLineEntry = PluginDefinition.findByName(previousLine[getAddressLine() - 2]).getLocaleEntry(locale);
 		cityEntry = PluginDefinition.findByName("CITY").getLocaleEntry(locale);
+		country = locale.getCountry().toUpperCase(Locale.ROOT);
+
+
+		if ("US".equals(country)) {
+			logicalZip = LogicalTypeFactory.newInstance(PluginDefinition.findByName("POSTAL_CODE.ZIP5_US"), analysisConfig);
+			logicalState = LogicalTypeFactory.newInstance(PluginDefinition.findByName("STATE_PROVINCE.STATE_US"), analysisConfig);
+		}
 
 		return true;
 	}
@@ -101,6 +119,8 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 
 	private boolean validation(final String trimmedUpper, final boolean detectMode) {
 		final List<String> words = wordProcessor.asWords(trimmedUpper);
+		boolean postalFound = false;
+		boolean stateFound = false;
 
 		final int wordCount = words.size();
 		for (int i = 0; i < wordCount; i++) {
@@ -108,6 +128,12 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 			if (i != 0 && addressMarkers.contains(word))
 				return true;
 			else if (AddressCommon.isModifier(word, i == wordCount - 1) && words.size() != 1)
+				return true;
+			if (logicalZip != null && !postalFound && logicalZip.isValid(word))
+				postalFound = true;
+			if (logicalState != null && !stateFound && logicalState.isValid(word))
+				stateFound = true;
+			if (postalFound && stateFound)
 				return true;
 		}
 
@@ -148,20 +174,20 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 			return 0;
 
 		final char lastChar = dataStreamName.charAt(dataStreamName.length() - 1);
-		// If it looks like an address and the last character is a 2 so we are feeling really good
-		if (headerConfidence >= 90 && getIndicator() == '2' && lastChar == '2')
+		// If it looks like an address and the last character matches the address line we are looking for we feel really good
+		if (headerConfidence >= 90 && getAddressLine() == lastChar)
 			return 99;
 
 		// If this header is the same as the previous but with the previous number then we are pretty confident
-		final int index = previousStreamName.indexOf(getIndicator() - 1);
-		if (index != -1 && dataStreamName.equals(previousStreamName.substring(0, index) + getIndicator() + previousStreamName.substring(index + 1)))
+		final int index = previousStreamName.indexOf(getAddressLineChar() - 1);
+		if (index != -1 && dataStreamName.equals(previousStreamName.substring(0, index) + getAddressLineChar() + previousStreamName.substring(index + 1)))
 			return headerConfidence >= 90 ? 99 : 85;
 
-		if (headerConfidence == 0 && lastChar != getIndicator())
+		if (headerConfidence == 0 && lastChar != getAddressLineChar())
 			return 0;
 
-		// Does the previous field look like an Address Line 1 AND not look like an Address Line 2
-		if (addressLine1Entry.getHeaderConfidence(previousStreamName) >= 90 && getHeaderConfidence(previousStreamName) < 99) {
+		// Does the previous field look like an Address Line <N-1> AND not look like an Address Line <N>
+		if (previousAddressLineEntry.getHeaderConfidence(previousStreamName) >= 90 && getHeaderConfidence(previousStreamName) < 99) {
 			if (current + 1 < context.getCompositeStreamNames().length &&
 					cityEntry.getHeaderConfidence(context.getCompositeStreamNames()[current + 1]) > 0)
 				return headerConfidence == 0 ? 85 : 95;
@@ -173,20 +199,20 @@ public abstract class AddressLineNEN extends LogicalTypeInfinite {
 
 	@Override
 	public double getConfidence(final long matchCount, final long realSamples, final AnalyzerContext context) {
-		final int confidence = headerConfidence(context);
+		final int headerConfidence = headerConfidence(context);
 
 		// If we really don't like the header call it a day
-		if (confidence == 0)
+		if (headerConfidence == 0)
 			return 0.0;
 
 		final double matchConfidence = (double)matchCount/realSamples;
 
 		// If we are super confident in the header boost if we like the data
-		if (confidence == 99)
+		if (headerConfidence == 99)
 			return matchConfidence > .25 ? .95 : .9;
 
 		// Header is reasonable - so check we have some data that looks reasonable
-		if (confidence >= 90)
+		if (headerConfidence >= 90)
 			return matchConfidence > .25 ? .9 : 0;
 
 		return matchConfidence > .5 ? .9 : 0;